+++ private/bufpkg/bufcheck/bufbreaking/bufbreaking_test.go
@@ -16,7 +16,10 @@ package bufbreaking_test
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"path/filepath"
+	"strings"
 	"testing"
 	"time"
 
@@ -1335,7 +1338,9 @@ func testBreaking(
 	require.NoError(t, err)
 	image = bufimage.ImageWithoutImports(image)
 
-	breakingConfig := workspace.GetBreakingConfigForOpaqueID(".")
+	opaqueID, err := testGetRootOpaqueID(workspace)
+	require.NoError(t, err)
+	breakingConfig := workspace.GetBreakingConfigForOpaqueID(opaqueID)
 	require.NotNil(t, breakingConfig)
 	handler := bufbreaking.NewHandler(
 		zap.NewNop(),
@@ -1357,5 +1362,22 @@ func testBreaking(
 			expectedFileAnnotations,
 			fileAnnotationSet.FileAnnotations(),
 		)
+	}
+}
+
+func testGetRootOpaqueID(workspace bufworkspace.Workspace) (string, error) {
+	var rootModules []bufmodule.Module
+	for _, module := range workspace.Modules() {
+		if strings.HasPrefix(module.OpaqueID(), ".") {
+			rootModules = append(rootModules, module)
+		}
+	}
+	switch len(rootModules) {
+	case 0:
+		return "", errors.New("no module with opaque ID starting with \".\"")
+	case 1:
+		return rootModules[0].OpaqueID(), nil
+	default:
+		return "", fmt.Errorf("got %d modules with opaque ID starting with \".\"", len(rootModules))
 	}
 }