+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [Unreleased]
 
 - Fix git input handling of relative HEAD refs without branch names.
+- Add `includes` key to module configurations in v2 `buf.yaml`, accepting a list of directories.
+  * If `includes` is specified, a proto file is considered in the module only if it is in one of the
+    directories specified.
+  * If both `includes` and `excludes` keys are specified for a module, a proto file is considered
+    part of this module if it is contained in any of the include paths and not in any of the exclude
+    paths.
+- Allow multiple module configurations in the same v2 `buf.yaml` to have the same directory path.
 
 ## [v1.38.0] - 2024-08-22